feat: add single-kanji database query from CLI
This adds a new CLI option to the db action to query a single kanji by
literal. The feature is implemented in a new `db/transact.py` module.
This renames `db/db.py` to `db/main.py` and also adds `__str__`
implementations to `Reading` and `Kanji` to facilitate human-friendlier
output to the console.