feat: fix Kanji interface to persistent database and add tests
Kanji instances were being created with integer radicals although the
member type was annotated as Radical. To fix this, Radical instances now
keep track of their own official radical number so that it can be used
when converting a Kanji instance to a database row as the database
expects the radical to be an int, not a serialized Radical. This change
also adds tests for the new Radical feature.

The implementation of libjyk.jykdb.build was moved into _build so that
it can be tested without requiring the full set of jouyou kanji. The
public build function retains the jouyou-kanji-related assert.